Iterate through attendees
The last section you’ll add to the detail screen is the attendees list.
ForEach is a ViewBuilder that computes views on demand from an
underlying collection of data. In this section, you’ll use ForEach to
compose a sequence of text views by iterating through a list of
attendees in the sample data.