Pass a binding into the detail view
The list view also needs to reflect changes a user makes to an
individual scrum on the edit screen. For this data flow, you need to
pass another binding down the view hierarchy.

You’ll start by changing the scrums property from a constant to a
binding. Then, you’ll create bindings from individual items in the
scrums array to the detail view.